post recieve hook and blob content invalidation works

+++ src/thruflo/model.py
@@ -307,7 +307,6 @@ class Blob(BaseDocument):
             path=path
         )
     
-    
     def update_data(self, github, save=True):
         """This is where the magic happens.
           
@@ -359,10 +358,7 @@ class Blob(BaseDocument):
         return changed
         
     
-    
     def get_data(self, github):
-        logging.debug('Blob.get_data: %s' % self.path)
-        
         self.update_data(github)
         return self.data
         
@@ -662,40 +658,28 @@ class Document(SluggedDocument):
     
     blobs = StringListProperty()
     
-    def get_blobs(self, github):
+    def get_blobs(self, github, force_update=True):
+        """
+        """
+        
         blobs = Blob.view(
             '_all_docs', 
             keys=self.blobs,
             include_docs=True
         ).all()
         
-        logging.info('get blobs')
-        
-        logging.info('before')
-        logging.info([blob.to_json() for blob in blobs])
-        
-        to_save = {}
-        for item in blobs:
-            # if it's been updated once, use the updated version
-            if to_save.has_key(item.id):
-                item = to_save.get(item.id)
-            elif item.update_data(github, save=False):
-                to_save[item.id] = item
-        if to_save:
-            dicts = [item.to_json() for item in to_save.values()]
-            Blob.get_db().bulk_save(dicts)
-        
-        logging.info('after')
-        logging.info([blob.to_json() for blob in blobs])
-        
-        logging.info('after a re-fetch')
-        blobs = Blob.view(
-            '_all_docs', 
-            keys=self.blobs,
-            include_docs=True
-        ).all()
-        logging.info([blob.to_json() for blob in blobs])
-        
+        if force_update:
+            to_save = {}
+            for item in blobs:
+                # if it's been updated once, use the updated version
+                if to_save.has_key(item.id):
+                    item = to_save.get(item.id)
+                elif item.update_data(github, save=False):
+                    to_save[item.id] = item
+            if to_save:
+                dicts = [item.to_json() for item in to_save.values()]
+                Blob.get_db().bulk_save(dicts)
+            
         return blobs
